pyproject.toml
[project]
name = "ai-news-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
]
